requirements.txt
# Project Dependencies
# Install with: pip install -r requirements.txt

# Core Data & ML
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Deep Learning & Computer Vision
torch>=1.10.0
torchvision>=0.11.0
timm>=0.9.0          # EfficientNet-B3 and 1000+ pretrained models
tensorflow>=2.8.0
# Note: tensorflow and torch can both be installed, or choose one per your preference

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Model & Data Management
h5py>=3.4.0
tqdm>=4.62.0

# API & Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Testing & Validation
pytest>=6.2.4
pytest-cov>=2.12.1

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipython>=7.25.0

